\name{exampleData}
\docType{data}
\alias{exampleData}
\alias{raw.dat}
\alias{longi.prime.dat}
\alias{longi.dat}
\alias{cart.dat}
\title{A small data set to use in function examples}
\description{
Imaging  data for 20 of the plants from an experiment in a Smarthouse in the Plant Accelerator. It is used as a small example in the documentation for \code{growthPheno}.
}
\usage{data(exampleData)}
\format{Four \code{data.frames}: raw.dat (280 rows by 33 columns), longi.prime.dat (280 rows by 45 columns), longi.dat (280 rows by 63 columns), cart.dat (20 rows by 14 columns).}
\keyword{datasets}