\name{growthPheno-deprecated}
\alias{growthPheno-deprecated}
\alias{getDates}
\title{Deprecated Functions in the Package asremlPlus}
\description{These functions have been renamed and deprecated in \code{growthPheno}: 
    \enumerate{
      \item getDates -> \code{\link{getTimesSubset}}
}} 
\usage{getDates(...)
}
\arguments{
  \item{...}{absorbs arguments passed from the old functions of the style foo.bar().}
}
\author{Chris Brien}
