\name{growthPheno-deprecated}
\alias{growthPheno-deprecated}
\alias{getDates}
\alias{anomPlot}
\alias{corrPlot}
\alias{imagetimesPlot}
\alias{longiPlot}
\title{Deprecated Functions in the Package asremlPlus}
\description{These functions have been renamed and deprecated in \code{growthPheno}: 
    \enumerate{
      \item getDates -> \code{\link{getTimesSubset}}
      \item anomPlot -> \code{\link{plotAnom}}
      \item corrPlot -> \code{\link{plotCorrmatrix}}
      \item imagetimesPlot -> \code{\link{plotImagetimes}}
      \item longiPlot -> \code{\link{plotLongitudinal}}
}} 
\usage{getDates(...)
anomPlot(...)
corrPlot(...)
imagetimesPlot(...)
longiPlot(...)
}
\arguments{
  \item{...}{absorbs arguments passed from the old functions of the style foo.bar().}
}
\author{Chris Brien}
