\name{RiceRaw.dat}
\docType{data}
\alias{RiceRaw.dat}
\title{Data for an experiment to investigate a rice germplasm panel}
\description{
The data is from an experiment in a Smarthouse in the Plant Accelerator. It is described in Al-Tamimi et al. (2016). It is used in \code{\link{growthPheno-pkg}} as an executable example to illustrate the use of \code{growthPheno}.

}
\usage{data(RiceRaw.dat)}
\format{A data.frame containing 1120 observations on 15 variables.}
\references{
Al-Tamimi, N, Brien, C.J., Oakey, H., Berger, B., Saade, S., Ho, Y. S., Schmockel, S. M., Tester, M. and Negrao, S. (2016) New salinity tolerance loci revealed in rice using high-throughput non-invasive phenotyping. \emph{Nature Communications}, \bold{7}, 13342. Retrieved from \doi{10.1038/ncomms13342}.
}
\source{Al-Tamimi N, Brien C, Oakey H, Berger B, Saade S, Ho YS, Schmockel SM, Tester M, Negrao S: Data from: Salinity tolerance loci revealed in rice using high-throughput non-invasive phenotyping. Retrieved from: \doi{10.5061/dryad.3118j}.}
\keyword{datasets}